Pin dycore, physics, and driver GFDL submodules to latest `main` (#11)
This PR updates the dycore, physics, and driver NOAA-GFDL submodules in this repo to their latest `main` branches.  Now that

- https://github.com/NOAA-GFDL/SHiELD_physics/pull/31
- https://github.com/NOAA-GFDL/atmos_drivers/pull/31
- https://github.com/NOAA-GFDL/atmos_drivers/pull/30
- https://github.com/NOAA-GFDL/GFDL_atmos_cubed_sphere/pull/302

have been merged, all the required features in SHiELD are in place to bring SHiELD-wrapper up-to-par with fv3gfs-wrapper.

Note that https://github.com/NOAA-GFDL/SHiELD_physics/pull/31 made changes to the long names of some of the diagnostics output in our regression test, which required a checksum update.  

I have also taken the liberty to remove the disclaimer from the README that states that changes to SHiELD are required to bring SHiELD-wrapper up-to-par with FV3GFS.  That is no longer a limiting factor, and those remaining features are implemented in open PRs in this repo.