Port tests for prescribing the sea surface temperature (#6)
In SHiELD this seems to be relatively straightforward. The one wrinkle, which tripped me up initially, was whether you are running with the mixed layer ocean model active or not. We will punt on that case for now, but it seems possible to set the surface temperature the ocean is restored to in that context, which we could consider adding appropriate getters, setters, and tests for in the future.

Note relative to FV3GFS we switch to prescribing a constant quasi-realistic value which is less jarring than prescribing random SSTs between 0 and 1 K.