Build in `repro` mode with `AVX_LEVEL=-mavx` (#12)
This PR modifies the build and test system to assume repro build mode and AVX_LEVEL=-mavx. This is consistent with how we build FV3GFS, and leads to informative tracebacks when the model crashes, and docker image builds that can run on varied hardware. Updating the AVX_LEVEL required updating the regression test checksums. I also took the liberty to generalize the wrapper Makefile to support different compile modes.

Addresses #9
Addresses #10