Update forcing data used in test configs (#20)
This PR updates the forcing data used in the test configs to point to data in the more conventional `vcm-fv3config` bucket instead of the `vcm-ml-experiments` bucket.  It also updates the configs to use tiled surface forcing files where available.  

Note this changes answers, so the regression test checksums need to be updated.